fixed extractText bugs
