+++ dfgfdgcutwav-meta_1601046535249.json
@@ -1,0 +1,26 @@
+  ExifToolVersion = 11.26
+  FileName = dfgfdgcut.wav
+  Directory = storage/convertedFiles
+  FileSize = 1764142
+  FileModifyDate = 1601046535.45994
+  FileAccessDate = 1601046535.45994
+  FileCreateDate = 1601046523.48339
+  FilePermissions = 33206
+  FileType = WAV
+  FileTypeExtension = WAV
+  MIMEType = audio/x-wav
+RIFF 'fmt ' chunk (16 bytes of data):
+  AudioFormat (SubDirectory) -->
+  + [BinaryData directory, 16 bytes]
+  | Encoding = 1
+  | NumChannels = 2
+  | SampleRate = 44100
+  | AvgBytesPerSec = 176400
+  | BitsPerSample = 16
+RIFF 'LIST_INFO' chunk (86 bytes of data):
+  Info (SubDirectory) -->
+  + [Info directory, 86 bytes]
+  | Artist = Carlos Vives & Sebastian Yatra
+  | Title = Robarte Un Beso
+  | Software = Lavf58.51.101
+RIFF 'data' chunk (1764000 bytes of data):
