+++ src/main/java/org/fundacionjala/converter/model/command/extractText/ExtractTextFacade.java
@@ -8,6 +8,7 @@
  */
 package org.fundacionjala.converter.model.command.extractText;
 
+import com.lowagie.text.DocumentException;
 import org.fundacionjala.converter.executor.Executor;
 import org.fundacionjala.converter.model.ChecksumMD5;
 import org.fundacionjala.converter.model.commons.exception.InvalidDataException;
@@ -15,10 +16,15 @@ import org.fundacionjala.converter.model.commons.validation.FormatValidation;
 import org.fundacionjala.converter.model.configPath.ConfigPath;
 import org.fundacionjala.converter.model.parameter.extractText.ExtractTextParameter;
 import org.fundacionjala.converter.model.utility.ConvertDoc;
+
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 /**
  * @author Angela Martinez
@@ -47,7 +53,7 @@ public class ExtractTextFacade {
      * Creates a document
      * @param parameter - the parameter to execute the conversion using tesseract
      */
-    public List<String> extractText(final ExtractTextParameter parameter) throws InvalidDataException, Exception {
+    public List<String> extractText(final ExtractTextParameter parameter) throws InvalidDataException, InterruptedException, ExecutionException, DocumentException, NoSuchAlgorithmException, IOException {
         parameter.validate();
         String format = parameter.getFormat();
         parameter.setOutputFile(configPath.getConvertedFilesPath());
@@ -79,7 +85,7 @@ public class ExtractTextFacade {
      * @param format
      * @param outputFile
      */
-    private void eraseDocument(final String format, final String outputFile) throws Exception {
+    private void eraseDocument(final String format, final String outputFile) throws IOException {
         if (format.equals(FormatValidation.FORMAT_DOCX) || format.equals(FormatValidation.FORMAT_PDF) || format.equals(FormatValidation.FORMAT_TXT)) {
             Files.delete(Paths.get(outputFile + FormatValidation.FORMAT_TXT));
         }
