+++ src/main/java/org/fundacionjala/converter/model/command/extractText/ReaderText.java
@@ -9,7 +9,9 @@
 package org.fundacionjala.converter.model.command.extractText;
 
 import java.io.BufferedReader;
+import java.io.FileNotFoundException;
 import java.io.FileReader;
+import java.io.IOException;
 
 public class ReaderText {
     private BufferedReader buffer;
@@ -22,23 +24,17 @@ public class ReaderText {
      * @param fileName the path and name of the file
      * @return the content of the file in a string
      */
-    public String readFile(final String fileName) throws Exception {
+    public String readFile(final String fileName) throws IOException {
         String readString = "";
         try {
-            String sCurrentLine;
+            String currentLine;
             buffer = new BufferedReader(new FileReader(fileName));
-            while ((sCurrentLine = buffer.readLine()) != null) {
-                readString += sCurrentLine;
+            while ((currentLine = buffer.readLine()) != null) {
+                readString += currentLine;
             }
-        } catch (Exception e) {
-            throw e;
         } finally {
-            try {
-                if (buffer != null) {
-                    buffer.close();
-                }
-            } catch (Exception ex) {
-                throw ex;
+            if (buffer != null) {
+                buffer.close();
             }
         }
         return readString;
