+++ src/main/resources/templates/audio/audio.html
@@ -82,8 +82,8 @@
         <div class="col s3 right">
             <form method="get" th:action="@{/download/}">
                 <input type="text" name="path" id="path" value="" style="display:none">
-                <input type="text" name="filenameDownload" id="filenameDownload" value="" style="display:none">
                 <button type="submit" id="submited" class="btn waves-effect waves-light" disabled>
+                    <input type="text" name="filenameDownload" id="filenameDownload" value="" style="display:none">
                     <i class="material-icons right">file_download</i>
                     Download File</button>
             </form>
