+++ src/main/resources/templates/extract/extractText.html
@@ -42,7 +42,7 @@
                     </select>
                     </div>
                     <div class="input-field col s12 m6">
-                        <input type="text" id="md5" name="md5" class="validate">
+                        <input type="text" id="md5" name="md5" class="validate" readonly="readonly">
                         <label for="md5">MD5:</label>
                     </div>
